Moved data format conversion to bean and created playback graph

+++ src/main/java/soundanalyzer/audio/AudioDataListener.java
@@ -1,5 +1,5 @@
 package soundanalyzer.audio;
 
 public interface AudioDataListener {
-    public void readData(double[] data);
+    public void readData(byte[] data);
 }
